Move from_reflect_or_world to common reflect module
